+++ kibot/config_reader.py
@@ -327,6 +327,7 @@ class CfgYamlReader(object):
         data = self.load_yaml(fstream)
         # Transfer command line global overwrites
         GS.global_output = GS.global_from_cli.get('output', None)
+        GS.global_dir = GS.global_from_cli.get('dir', None)
         GS.global_variant = GS.global_from_cli.get('variant', None)
         GS.global_kiauto_wait_start = GS.global_from_cli.get('kiauto_wait_start', None)
         GS.global_kiauto_time_out_scale = GS.global_from_cli.get('kiauto_time_out_scale', None)